git-svn-id: https://svn.kuali.org/repos/foundation/trunk@22217 eb505253-de9c-4f3b-bf87-3aad84f29c0f

+++ kuali-util/src/main/java/org/kuali/common/util/LongCounter.java
@@ -34,13 +34,13 @@ public class LongCounter {
 
 	public synchronized long increment(long amount) {
 		Assert.isTrue(amount > 0, "amount must be > 0");
-		Assert.isTrue(count <= Integer.MAX_VALUE - amount, "max value exceeded");
+		Assert.isTrue(count <= Long.MAX_VALUE - amount, "max value exceeded");
 		return count += amount;
 	}
 
 	public synchronized long decrement(long amount) {
 		Assert.isTrue(amount > 0, "amount must be > 0");
-		Assert.isTrue(count >= Integer.MIN_VALUE + amount, "min value exceeded");
+		Assert.isTrue(count >= Long.MIN_VALUE + amount, "min value exceeded");
 		return count -= amount;
 	}
